+++ src/main/java/com/WG/WithGoods/service/LimitedProductService.java
@@ -32,8 +32,8 @@ public class LimitedProductService {
         return toDto(limitedProductRepository.save(limited));
     }
 
-    public LimitedProductDto get(Long id) {
-        return toDto(limitedProductRepository.findById(Math.toIntExact(id))
+    public LimitedProductDto get(Integer id) {
+        return toDto(limitedProductRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 한정판 상품")));
     }
 
@@ -41,8 +41,8 @@ public class LimitedProductService {
         return limitedProductRepository.findAll().stream().map(this::toDto).toList();
     }
 
-    public LimitedProductDto update(Long id, LimitedProductDto dto) {
-        LimitedProduct limited = limitedProductRepository.findById(Math.toIntExact(id))
+    public LimitedProductDto update(Integer id, LimitedProductDto dto) {
+        LimitedProduct limited = limitedProductRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("수정할 한정판 상품이 없습니다."));
 
         limited.setStartDate(dto.getStartDate());
@@ -52,8 +52,8 @@ public class LimitedProductService {
         return toDto(limitedProductRepository.save(limited));
     }
 
-    public void delete(Long id) {
-        limitedProductRepository.deleteById(Math.toIntExact(id));
+    public void delete(Integer id) {
+        limitedProductRepository.deleteById(id);
     }
 
     private LimitedProductDto toDto(LimitedProduct entity) {
